requirements.txt
streamlit
faiss-cpu
numpy
pickle-mixin
requests
sentence-transformers
pdfplumber
